# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Employee {
  id: number;
  name: string;
  role: string;
  department: string;
  avatarInitials?: string;
  color?: string;
}

export interface EmployeeInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  /** @minLength 1 */
  department: string;
  color?: string;
}

export interface Meeting {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  startTime: string;
  endTime: string;
  organizerId: number;
  organizer?: Employee;
  participants: Employee[];
  /** @nullable */
  location?: string | null;
  createdAt: string;
}

export interface MeetingInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  startTime: string;
  endTime: string;
  organizerId: number;
  /** @minItems 1 */
  participantIds: number[];
  location?: string;
}

export interface ConflictCheckInput {
  startTime: string;
  endTime: string;
  participantIds: number[];
  /** @nullable */
  excludeMeetingId?: number | null;
}

export type ConflictCheckResultConflictsItem = {
  employeeId: number;
  employeeName: string;
  conflictingMeetingId: number;
  conflictingMeetingTitle: string;
};

export interface ConflictCheckResult {
  hasConflict: boolean;
  conflicts: ConflictCheckResultConflictsItem[];
}

export type ConflictErrorConflictsItem = {
  employeeId: number;
  employeeName: string;
  conflictingMeetingId: number;
  conflictingMeetingTitle: string;
};

export interface ConflictError {
  error: string;
  conflicts: ConflictErrorConflictsItem[];
}

export interface ErrorResponse {
  error: string;
}

export type ListMeetingsParams = {
/**
 * Start date (YYYY-MM-DD)
 */
from?: string;
/**
 * End date (YYYY-MM-DD)
 */
to?: string;
/**
 * Filter by employee ID
 * @nullable
 */
employeeId?: number | null;
};