# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meeting calendar API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: employees
    description: Employee operations
  - name: meetings
    description: Meeting operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List meetings by date range
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
          description: Start date (YYYY-MM-DD)
        - in: query
          name: to
          schema:
            type: string
            format: date
          description: End date (YYYY-MM-DD)
        - in: query
          name: employeeId
          schema:
            type: ["integer", "null"]
          description: Filter by employee ID
      responses:
        "200":
          description: List of meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
    post:
      operationId: createMeeting
      tags: [meetings]
      summary: Create a new meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingInput"
      responses:
        "201":
          description: Created meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "400":
          description: Bad request or conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Scheduling conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"

  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get a meeting by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMeeting
      tags: [meetings]
      summary: Delete a meeting
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /meetings/today:
    get:
      operationId: getTodayMeetings
      tags: [meetings]
      summary: Get all meetings for today
      responses:
        "200":
          description: Today meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"

  /meetings/week:
    get:
      operationId: getWeekMeetings
      tags: [meetings]
      summary: Get all meetings for current week
      responses:
        "200":
          description: Week meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"

  /meetings/check-conflict:
    post:
      operationId: checkConflict
      tags: [meetings]
      summary: Check if a time slot has conflicts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConflictCheckInput"
      responses:
        "200":
          description: Conflict check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictCheckResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Employee:
      type: object
      required: [id, name, role, department]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        department:
          type: string
        avatarInitials:
          type: string
        color:
          type: string

    EmployeeInput:
      type: object
      required: [name, role, department]
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
        department:
          type: string
          minLength: 1
        color:
          type: string

    Meeting:
      type: object
      required: [id, title, startTime, endTime, organizerId, participants, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        organizerId:
          type: integer
        organizer:
          $ref: "#/components/schemas/Employee"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        location:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    MeetingInput:
      type: object
      required: [title, startTime, endTime, organizerId, participantIds]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        organizerId:
          type: integer
        participantIds:
          type: array
          items:
            type: integer
          minItems: 1
        location:
          type: string

    ConflictCheckInput:
      type: object
      required: [startTime, endTime, participantIds]
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        participantIds:
          type: array
          items:
            type: integer
        excludeMeetingId:
          type: ["integer", "null"]

    ConflictCheckResult:
      type: object
      required: [hasConflict, conflicts]
      properties:
        hasConflict:
          type: boolean
        conflicts:
          type: array
          items:
            type: object
            required: [employeeId, employeeName, conflictingMeetingId, conflictingMeetingTitle]
            properties:
              employeeId:
                type: integer
              employeeName:
                type: string
              conflictingMeetingId:
                type: integer
              conflictingMeetingTitle:
                type: string

    ConflictError:
      type: object
      required: [error, conflicts]
      properties:
        error:
          type: string
        conflicts:
          type: array
          items:
            type: object
            required: [employeeId, employeeName, conflictingMeetingId, conflictingMeetingTitle]
            properties:
              employeeId:
                type: integer
              employeeName:
                type: string
              conflictingMeetingId:
                type: integer
              conflictingMeetingTitle:
                type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string