# lib/api-zod/src/generated/types/conflictCheckResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.1.0
 */
import type { ConflictCheckResultConflictsItem } from './conflictCheckResultConflictsItem';

export interface ConflictCheckResult {
  hasConflict: boolean;
  conflicts: ConflictCheckResultConflictsItem[];
}