# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  name: string;
  role: string;
  department: string;
  avatarInitials?: string;
  color?: string;
}