# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  /** @minLength 1 */
  department: string;
  color?: string;
}