# lib/api-zod/src/generated/types/meeting.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.1.0
 */
import type { Employee } from './employee';

export interface Meeting {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  startTime: Date;
  endTime: Date;
  organizerId: number;
  organizer?: Employee;
  participants: Employee[];
  /** @nullable */
  location?: string | null;
  createdAt: Date;
}