# lib/api-zod/src/generated/types/meetingInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.1.0
 */

export interface MeetingInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  startTime: Date;
  endTime: Date;
  organizerId: number;
  /** @minItems 1 */
  participantIds: number[];
  location?: string;
}